int to bigint

+++ models.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, ForeignKey, Integer, String
+from sqlalchemy import Column, ForeignKey, Integer, String, BigInteger
 from sqlalchemy.ext.declarative import declarative_base
 
 
@@ -6,7 +6,7 @@ Base = declarative_base()
 
 
 class Rooms(Base):
-    __tablename__ = 'rooms'
+    __tablename__ = "rooms"
     id = Column(Integer, primary_key=True, autoincrement=True)
     creator_id = Column(Integer, ForeignKey("users.id"))
     name = Column(String, nullable=False)
@@ -15,8 +15,8 @@ class Rooms(Base):
 
 
 class Users(Base):
-    __tablename__ = 'users'
-    id = Column(Integer, primary_key=True)
+    __tablename__ = "users"
+    id = Column(BigInteger, primary_key=True)
     username = Column(String, nullable=False)
     first_name = Column(String, nullable=True)
     last_name = Column(String, nullable=True)
@@ -26,7 +26,7 @@ class Users(Base):
 
 
 class Pairs(Base):
-    __tablename__ = 'pairs'
+    __tablename__ = "pairs"
     giver_id = Column(Integer, ForeignKey("users.id"), primary_key=True)
     room_id = Column(Integer, ForeignKey("rooms.id"), primary_key=True)
     taker_id = Column(Integer, ForeignKey("users.id"))
